+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Send `User-Agent` information with metrics.
+
 ## [1.2.0] - 2022-02-01
 
 ### Added